smart_document_processor/requirements.txt
# Smart Document Processor Requirements
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==10.1.0
opencv-python==4.8.1.78
reportlab==4.0.7
openai==1.3.8
anthropic==0.7.8
jinja2==3.1.2
aiofiles==23.2.1
python-magic==0.4.27
websockets==12.0
pydantic==2.5.0
httpx==0.25.2
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6